requirements.txt
django>=1.5.1
django-tastypie>=0.9.12
psycopg2